v 1.0.0 배포 완료

+++ services/comment.service.js
@@ -1,6 +1,6 @@
 const CommentRepository = require('../repositories/comment.repository');
 const { Comment, Quiz, Member, CommentLike } = require('../models');
-const { NotFound } = require('../helper/http.exception.helper');
+const { NotFound, Forbidden } = require('../helper/http.exception.helper');
 
 class CommentService {
   commentRepository = new CommentRepository(Comment, Quiz, Member, CommentLike);
@@ -9,13 +9,41 @@ class CommentService {
     return await this.commentRepository.createComment(mId, qId, comment);
   };
 
-  getAllComments = async (qId) => {
+  getAllComments = async (qId, mId) => {
     const allComments = await this.commentRepository.getAllComments(qId);
+    const likedComment = await this.commentRepository.findLikeStatus(mId);
+
+    for (let i = 0; i < allComments.length; i++) {
+      allComments[i].isLikedComment = false;
+      allComments[i].isDisLikedComment = false;
+    }
+
+    for (let i = 0; i < allComments.length; i++) {
+      for (let j = 0; j < likedComment.length; j++) {
+        if (
+          allComments[i].cId === likedComment[j].cId &&
+          likedComment[j].commentLikeStatus === true
+        ) {
+          allComments[i].isLikedComment = true;
+        } else if (
+          allComments[i].cId === likedComment[j].cId &&
+          likedComment[j].commentLikeStatus === false
+        ) {
+          allComments[i].isDisLikedComment = true;
+        }
+      }
+    }
 
     return allComments;
   };
 
-  updateComment = async (cId, comment) => {
+  updateComment = async (cId, comment, mId) => {
+    const commentData = await this.commentRepository.findComment(cId);
+
+    if (!commentData) throw new NotFound('');
+
+    if (mId !== commentData.mId) throw new Forbidden('');
+
     const updateCommentData = await this.commentRepository.updateComment(
       cId,
       comment
@@ -26,7 +54,13 @@ class CommentService {
     }
   };
 
-  deleteComment = async (cId) => {
+  deleteComment = async (cId, mId) => {
+    const commentData = await this.commentRepository.findComment(cId);
+
+    if (!commentData) throw new NotFound('');
+
+    if (mId !== commentData.mId) throw new Forbidden('');
+
     const deleteCommentData = await this.commentRepository.deleteComment(cId);
 
     if (deleteCommentData < 1) {
